Add pagination and search functionality to getRestaurants method

+++ src/restaurant/restaurant.service.ts
@@ -1,10 +1,12 @@
 /* eslint-disable prettier/prettier */
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Restaurant, RestaurantDocument } from 'src/schemas/restaurant.schema';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
+import { Query } from 'express-serve-static-core';
+
 
 @Injectable()
 export class RestaurantService {
@@ -13,14 +15,31 @@ export class RestaurantService {
     private restaurantModel: Model<RestaurantDocument>,
   ) {}
 
-  async getRestaurants(): Promise<Restaurant[]> {
-    return this.restaurantModel.find().exec();
+  async getRestaurants(query: Query): Promise<Restaurant[]> {
+
+    const resPerPage = 5;
+    const currentPage = Number(query.page) || 1;
+    const skip = resPerPage * (currentPage - 1);
+
+
+
+    const keyword = query.keyword ? {
+      name: {
+        $regex: query.keyword,
+        $options: 'i',
+      },
+    } : {};
+    
+    const restaurants = await this.restaurantModel.find({ ...keyword }).populate("comments").limit(resPerPage).skip(skip).exec();
+
+    return restaurants;
   }
 
   async getRestaurant(id: string): Promise<Restaurant> {
-    const restaurant = await this.restaurantModel.findById(id).exec();
+    const restaurant = await this.restaurantModel.findById(id).populate('comments').exec();
+
     if (!restaurant) {
-      throw new Error('Restaurant not found');
+      throw new HttpException('Restaurant not found', HttpStatus.NOT_FOUND);
     }
 
     return restaurant;
@@ -35,21 +54,32 @@ export class RestaurantService {
 
   async updateRestaurant(
     id: string,
+    ownerId: string,
     updateRestaurantDto: UpdateRestaurantDto,
   ): Promise<Restaurant> {
-    const restaurant = await this.restaurantModel
-      .findByIdAndUpdate(id, updateRestaurantDto, { new: true })
-      .exec();
+    const restaurant = await this.restaurantModel.findById(id).exec();
     if (!restaurant) {
-      throw new Error('Restaurant not found');
+      throw new HttpException('Restaurant not found', HttpStatus.NOT_FOUND);
     }
-    return restaurant;
+
+    if (restaurant.ownerId !== ownerId) {
+      throw new HttpException('You can only update your own restaurants', HttpStatus.UNAUTHORIZED);
+    }
+
+
+    return await this.restaurantModel
+    .findByIdAndUpdate(id, updateRestaurantDto, { new: true })
+    .exec();;
   }
 
-  async removeRestaurant(id: string): Promise<Restaurant> {
+  async removeRestaurant(id: string, ownerId: string): Promise<Restaurant> {
     const tobeRemoved = await this.restaurantModel.findById(id).exec();
     if (!tobeRemoved) {
-      throw new Error('Restaurant not found');
+      throw new HttpException('Restaurant not found', HttpStatus.NOT_FOUND);
+    }
+
+    if (tobeRemoved.ownerId !== ownerId) {
+      throw new HttpException('You can only update your own restaurants', HttpStatus.UNAUTHORIZED)
     }
 
     await this.restaurantModel.findByIdAndDelete(id).exec();
